glsl shader added

+++ CADie/src/Platform/Windows/WindowsInput.cpp
@@ -30,7 +30,7 @@ namespace CADie {
 		double xpos, ypos;
 		glfwGetCursorPos(window, &xpos, &ypos);
 
-		return std::pair<float, float>(xpos, ypos);
+		return std::pair<float, float>((float)xpos, (float)ypos);
 	}
 
 	float WindowsInput::GetMouseXImpl()
